requirements.txt
openai==1.57.0
google-cloud-aiplatform==1.67.0
anthropic==0.40.0
PyPDF2==3.0.1
EbookLib==0.18
bs4==0.0.2
beautifulsoup4==4.12.3
protobuf==5.29.3
python-dotenv==1.0.1
google-generativeai==0.8.3
pymongo==4.11.1
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
numpy==2.2.1
xlsxwriter